라이팅에 대한 Manager의 버그 발생

+++ Portfolio1/Portfolio1/LightInterface.h
@@ -17,7 +17,7 @@ struct LightSet
 	float				fSpotPower;
 };
 
-enum ELightType : uint32_t
+enum ELightType : size_t
 {
 	PointLightType = 1,
 	SpotLightType
